requirements.txt
amqp==2.2.2
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.2.2
cssselect==1.0.3
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.0.3
django-cors-headers==2.4.0
django-filter==1.1.0
django-guardian==1.4.9
django-haystack==2.8.0
django-import-export==1.0.1
django-ipware==2.1.0
django-js-asset==1.0.0
django-mptt==0.9.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-redis==4.9.0
django-rest-swagger==2.1.2
django-simple-captcha==0.5.6
django-templated-mail==1.1.1
django-versatileimagefield==1.9
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djoser==1.2.0
drf-haystack==1.8.3
drf-writable-nested==0.4.2
dry-rest-permissions==0.1.10
elasticsearch==2.2.0
et-xmlfile==1.0.1
eventlet==0.22.1
greenlet==0.4.13
html2text==2018.1.9
httplib2==0.11.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
itypes==1.1.0
jdcal==1.4
jieba==0.39
Jinja2==2.10
kombu==4.1.0
lxml==4.2.2
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.12
oauthlib==2.0.7
odfpy==1.3.6
openapi-codec==1.3.2
openpyxl==2.5.4
parsel==1.4.0
Pillow==5.0.0
pipenv==2018.7.1
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycparser==2.18
pycryptodomex==3.6.6
PyDispatcher==2.0.5
PyJWT==1.6.1
pyOpenSSL==18.0.0
python-alipay-sdk==1.7.1
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.3
PyYAML==3.13
queuelib==1.5.0
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
rest-framework-generic-relations==1.2.0
Scrapy==1.5.0
scrapy-djangoitem==1.1.1
service-identity==17.0.0
simplejson==3.16.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
tablib==0.11.4
Twisted==18.4.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
uWSGI==2.0.17
vine==1.1.4
virtualenv==16.0.0
virtualenv-clone==0.3.0
w3lib==1.19.0
Whoosh==2.7.4
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.5.0
